requirements.txt
streamlit>=1.36
langchain>=0.2
langchain-core>=0.2
langchain-community>=0.2
langchain-openai>=0.2
pinecone-client==3.2.2
pypdf
pymupdf
